+++ tests/Integration/NotionExportIntegrationTests.cs
@@ -1,5 +1,5 @@
-using NotionMarkdownConverter.Application.Abstractions;
 using NotionMarkdownConverter.Extract;
+using NotionMarkdownConverter.Infrastructure;
 using NotionMarkdownConverter.Pipeline;
 using NotionMarkdownConverter.Pipeline.Models;
 using NotionMarkdownConverter.Transform;
