+++ tests/Unit/NotionPageTransformerTests.cs
@@ -1,8 +1,8 @@
 using Microsoft.Extensions.DependencyInjection;
 using Notion.Client;
 using NotionMarkdownConverter.Application;
-using NotionMarkdownConverter.Application.Abstractions;
 using NotionMarkdownConverter.Application.Configuration;
+using NotionMarkdownConverter.Infrastructure;
 using NotionMarkdownConverter.Pipeline.Models;
 using NotionMarkdownConverter.Shared.Enums;
 using NotionMarkdownConverter.Shared.Models;
